REFACTOR: risky - move functions to new module
